chore: finish splitting arrays/mod
